// src/common/common.types.ts
export const RoleType = {
  SUPERADMIN: 'SUPERADMIN',
  PATIENT: 'PATIENT',
  PHYSICIAN: 'PHYSICIAN',
  GUARDIAN: 'GUARDIAN',
  NURSE_PRACTITIONER: 'NURSE_PRACTITIONER',
  PHYSICIAN_ASSISTANT: 'PHYSICIAN_ASSISTANT',
  OFFICE_STAFF: 'OFFICE_STAFF',
  REGULAR_NURSE: 'REGULAR_NURSE',
  BILLING_STAFF: 'BILLING_STAFF',

  CHI_US_ADMIN: 'CHI_US_ADMIN',
  CHI_US_PROVIDER: 'CHI_US_PROVIDER',
  CHI_TECH: 'CHI_TECH',
  CHI_PK_PROVIDER: 'CHI_PK_PROVIDER',
  CHI_PK_NURSE: 'CHI_PK_NURSE',
} as const;

export type RoleType = keyof typeof RoleType;